+++ hata/discord/guild/__init__.py
@@ -1,5 +1,6 @@
 from .audit_logs import *
 from .discovery import *
+from .event_types import *
 from .flags import *
 from .guild import *
 from .preinstanced import *
@@ -12,6 +13,7 @@ from .widget import *
 __all__ = (
     *audit_logs.__all__,
     *discovery.__all__,
+    *event_types.__all__,
     *flags.__all__,
     *guild.__all__,
     *preinstanced.__all__,