+++ hata/discord/guild/preinstanced.py
@@ -1,9 +1,7 @@
-__all__ = ('AuditLogEvent', 'ContentFilterLevel', 'GuildFeature', 'MFA', 'MessageNotificationLevel', 'NsfwLevel',
-    'VerificationLevel', 'VerificationScreenStepType', 'VoiceRegion', )
+__all__ = ('AuditLogEvent', 'ContentFilterLevel', 'GuildFeature', 'GuildJoinRequestStatus', 'MFA',
+    'MessageNotificationLevel', 'NsfwLevel', 'VerificationLevel', 'VerificationScreenStepType', 'VoiceRegion', )
 
-import warnings
-
-from scarletio import export, class_property
+from scarletio import export
 from ..bases import PreinstancedBase, Preinstance as P
 
 
@@ -792,6 +790,8 @@ class VerificationScreenStepType(PreinstancedBase):
 
     Attributes
     ----------
+    name : `str`
+        The verification screen step type's name.
     value : `str`
         The Discord side identifier value of the verification step types.
     
@@ -802,16 +802,16 @@ class VerificationScreenStepType(PreinstancedBase):
     VALUE_TYPE : `type` = `str`
         The verification screen steps' values' type.
     DEFAULT_NAME : `str` = `''`
-        The default name of the verification screen step types.Verification screen step types have the
-        same value as name, so at their case it is not applicable.
+        The default name of the verification screen step types. New verification screen step types have their name
+        generated from their value, so it is not applicable for them.
     
     Every predefined verification screen step type can be accessed as class attribute as well:
     
-    +-----------------------+-------+
-    | Class attribute names | Value |
-    +=======================+=======+
-    | rules                 | TERMS |
-    +-----------------------+-------+
+    +-----------------------+-----------+-------+
+    | Class attribute names | Name      | Value |
+    +=======================+===========+=======+
+    | rules                 | rules     | TERMS |
+    +-----------------------+-----------+-------+
     """
     INSTANCES = {}
     VALUE_TYPE = str
@@ -836,7 +836,7 @@ class VerificationScreenStepType(PreinstancedBase):
         """
         self = object.__new__(cls)
         self.value = value
-        self.name = value
+        self.name = value.lower()
         self.INSTANCES[value] = self
         return self
     
@@ -845,3 +845,68 @@ class VerificationScreenStepType(PreinstancedBase):
         return f'{self.__class__.__name__}(value={self.value!r})'
     
     rules = P('TERMS', 'rules')
+
+
+class GuildJoinRequestStatus(PreinstancedBase):
+    """
+    Represents the status of a ``GuildJoinRequestUpdateEvent``.
+    
+    Attributes
+    ----------
+    name : `str`
+        The name of the guild join request status.
+    value : `str`
+        The Discord side identifier value of the guild join request status.
+    
+
+    Class Attributes
+    ----------------
+    INSTANCES : `dict` of (`int`, ``GuildJoinRequestStatus``) items
+        Stores the predefined ``GuildJoinRequestStatus`` instances.
+    VALUE_TYPE : `type` = `str`
+        The guild join request statuses' values' type.
+    DEFAULT_NAME : `str` = `''`
+        The default name of the guild join request statuses. Guild join request statuses have their name generated from
+        their value, so at their case it is not applicable.
+    
+    Every predefined guild join request status can be accessed as class attribute as well:
+    
+    +-----------------------+-----------+-----------+
+    | Class attribute names | Name      | Value     |
+    +=======================+===========+===========+
+    | approved              | approved  | APPROVED  |
+    +-----------------------+-----------+-----------+
+    """
+    
+    INSTANCES = {}
+    VALUE_TYPE = str
+    DEFAULT_NAME = ''
+    
+    __slots__ = ()
+    
+    @classmethod
+    def _from_value(cls, value):
+        """
+        Creates a new guild join request status from the given value.
+        
+        Parameters
+        ----------
+        value : `str`
+            The guild join request status's identifier value.
+        
+        Returns
+        -------
+        self : ``GuildJoinRequestStatus``
+            The guild join request status.
+        """
+        self = object.__new__(cls)
+        self.value = value
+        self.name = value.lower()
+        self.INSTANCES[value] = self
+        return self
+    
+    def __repr__(self):
+        """Returns the representation of the guild join request status."""
+        return f'{self.__class__.__name__}(value={self.value!r})'
+    
+    approved = P('APPROVED', 'approved')