+++ hata/discord/scheduled_event/event_types.py
@@ -1,5 +1,7 @@
 __all__ = ('ScheduledEventSubscribeEvent', 'ScheduledEventUnsubscribeEvent')
 
+from scarletio import copy_docs
+
 from ..bases import EventBase
 
 class ScheduledEventSubscribeEvent(EventBase):
@@ -19,7 +21,7 @@ class ScheduledEventSubscribeEvent(EventBase):
     
     def __new__(cls, data):
         """
-        Creates a new scheduled event subscribe from the given data.
+        Creates a new scheduled event (un)subscribe event from the given data.
         
         Parameters
         ----------
@@ -37,8 +39,8 @@ class ScheduledEventSubscribeEvent(EventBase):
         return self
     
     
+    @copy_docs(EventBase.__repr__)
     def __repr__(self):
-        """Returns the representation of the scheduled event subscribe event."""
         repr_parts = [
             '<',
             self.__class__.__name__,
@@ -54,23 +56,40 @@ class ScheduledEventSubscribeEvent(EventBase):
         return ''.join(repr_parts)
     
     
+    @copy_docs(EventBase.__len__)
     def __len__(self):
-        """Helper for unpacking if needed."""
         return 3
     
     
+    @copy_docs(EventBase.__iter__)
     def __iter__(self):
-        """
-        Unpacks the scheduled event subscribe event.
-        
-        This method is a generator.
-        """
         yield self.guild_id
         yield self.scheduled_event_id
         yield self.user_id
+    
+    
+    @copy_docs(EventBase.__eq__)
+    def __eq__(self, other):
+        if type(self) is not type(other):
+            return NotImplemented
+        
+        if self.guild_id != other.guild_id:
+            return False
+        
+        if self.scheduled_event_id != other.scheduled_event_id:
+            return False
+        
+        if self.user_id != other.user_id:
+            return False
+        
+        return True
+    
+    @copy_docs(EventBase.__hash__)
+    def __hash__(self):
+        return self.guild_id^self.scheduled_event_id^self.user_id
 
 
-class ScheduledEventUnsubscribeEvent(EventBase):
+class ScheduledEventUnsubscribeEvent(ScheduledEventSubscribeEvent):
     """
     Represents a `GUILD_SCHEDULED_EVENT_USER_REMOVE` event.
     
@@ -83,56 +102,4 @@ class ScheduledEventUnsubscribeEvent(EventBase):
     user_id : `int`
         The identifier of the user, who unsubscribed to the event.
     """
-    __slots__ = ('guild_id', 'scheduled_event_id', 'user_id', )
-    
-    def __new__(cls, data):
-        """
-        Creates a new scheduled event unsubscribe from the given data.
-        
-        Parameters
-        ----------
-        data : `dict` of (`str`, `Any`) items
-            Scheduled event unsubscribe event data.
-        """
-        guild_id = int(data['guild_id'])
-        scheduled_event_id = int(data['guild_scheduled_event_id'])
-        user_id = int(data['user_id'])
-        
-        self = object.__new__(cls)
-        self.guild_id = guild_id
-        self.scheduled_event_id = scheduled_event_id
-        self.user_id = user_id
-        return self
-    
-    
-    def __repr__(self):
-        """Returns the representation of the scheduled event unsubscribe event."""
-        repr_parts = [
-            '<',
-            self.__class__.__name__,
-            ' guild_id=',
-            repr(self.guild_id),
-            ', scheduled_event_id=',
-            repr(self.scheduled_event_id),
-            ', user_id=',
-            repr(self.scheduled_event_id),
-            '>'
-        ]
-        
-        return ''.join(repr_parts)
-    
-    
-    def __len__(self):
-        """Helper for unpacking if needed."""
-        return 3
-    
-    
-    def __iter__(self):
-        """
-        Unpacks the scheduled event unsubscribe event.
-        
-        This method is a generator.
-        """
-        yield self.guild_id
-        yield self.scheduled_event_id
-        yield self.user_id
+    __slots__ = ()